refactor: extract mcp auth hooks and utilities
