+++ frontend/src/metabase/embedding/mcp/auth/index.ts
@@ -1,0 +1,1 @@
+export { useMcpUiAuth } from "./useMcpUiAuth";
