--- frontend/src/metabase/embedding/mcp/utils/uiCredential.ts
@@ -1,36 +1,0 @@
-import { PLUGIN_API } from "metabase/api/client";
-
-import { MCP_APPS_METADATA_KEY } from "../constants";
-
-export interface McpUiAuth {
-  credential: string;
-  sessionId: string;
-}
-
-/**
- * Extract the UI credentials and session ID from tool result metadata.
- */
-export function getMcpUiAuthFromToolMetadata(
-  metadata: Record<string, unknown> | undefined,
-): McpUiAuth | null {
-  const value = metadata?.[MCP_APPS_METADATA_KEY];
-
-  if (typeof value !== "object" || value === null) {
-    return null;
-  }
-
-  const credential = "credential" in value ? value.credential : null;
-  const sessionId = "sessionId" in value ? value.sessionId : null;
-
-  if (typeof credential !== "string" || typeof sessionId !== "string") {
-    return null;
-  }
-
-  return { credential, sessionId };
-}
-
-export function installMcpUiCredential(credential: string) {
-  PLUGIN_API.onBeforeRequestHandlers.setEmbeddingRequestAuthHeaders =
-    // eslint-disable-next-line metabase/no-literal-metabase-strings -- request header name
-    async () => ({ headers: { "X-Metabase-Mcp-Ui-Auth": credential } });
-}
